pyproject.toml
[tool.poetry]
name = "two-factor-authentication"
version = "0.1.0"
description = ""
authors = ["vunk.lai <vunk.lai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
# django
django = "^4.2.1"
# other
pyotp = "^2.8.0"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.13.2"
# fomat
black = "^23.3.0"
isort = "^5.12.0"
# linting
pylint = "^2.17.4"
pylint-django = "^2.5.3"
mypy = "^1.3.0"

[tool.isort]
profile = "black"
known_django = "django,rest_framework"
sections = "FUTURE,STDLIB,DJANGO,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
skip = "migrations"

[tool.pylint.main]
load-plugins = [
  "pylint_django"
]
django-settings-module = "server.settings"
ignore = [
  "db.sqlite3",
]

[tool.pylint.messages_control]
disable = [
  "missing-module-docstring",
  "missing-class-docstring",
  "missing-function-docstring",
  "too-few-public-methods",
  "fixme",
]

[tool.pylint.basic]
good-names = [
  "pk",
  "_"
]

[tool.mypy]
ignore_missing_imports = true
exclude = [
  "/migrations/"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
